+++ src/app/processing/routines/hwindow.cpp
@@ -1,6 +1,6 @@
 #include "routines.h"
 
-void hwindow(const std::vector<int>& in, int ioff, std::vector<double>& out, int n,
+void hwindow(const std::vector<double>& in, int ioff, std::vector<double>& out, int n,
              double preEmphasis) {
     static std::vector<double> wind;
 
@@ -14,12 +14,11 @@ void hwindow(const std::vector<int>& in, int ioff, std::vector<double>& out, int
 
     if (preEmphasis != 0.) {
         for (int i = 0; i < n; ++i) {
-            out[i] =
-                wind[i] * (double(in[ioff + i + 1]) - preEmphasis * double(in[ioff + i]));
+            out[i] = wind[i] * (in[ioff + i + 1] - preEmphasis * in[ioff + i]);
         }
     } else {
         for (int i = 0; i < n; ++i) {
-            out[i] = wind[i] * double(in[ioff + i]);
+            out[i] = wind[i] * in[ioff + i];
         }
     }
 }
